// src/app/features/metrics/models/metrics-response.ts
// Generated by https://quicktype.io

export interface MetricsResponse {
  requestStatus: RequestStatus;
  totalStudents: number;
  laboratories: any[];
  subjects: Department;
  department: Department;
  careers: Careers;
}

export interface Careers {
  count: number;
  undergraduateLarge: Postgraduate;
  undergraduateShort: Postgraduate;
  postgraduate: Postgraduate;
  rows: any[];
}

export interface Postgraduate {
  count: number;
  totalStudents: number;
  totalApproved: number;
  totalRejected: number;
}

export interface Department {
  count: number;
  rows: any[];
}

export interface RequestStatus {
  status: string[];
  request: number[];
}